+++ crates/kct_compiler/src/extension/input.rs
@@ -1,11 +1,8 @@
-use crate::extension::{Extension, Name, Plugin};
+use crate::extension::{Extension, Name, Plugin, Property};
 use crate::{Input, Runtime};
 
 impl Extension for Input {
 	fn plug(&self, _: Runtime) -> Plugin {
-		Plugin::Property {
-			name: Name::Input,
-			value: self.0.clone(),
-		}
+		Plugin::Create(Property::Primitive(Name::Input, self.0.clone()))
 	}
 }