+++ crates/kct_compiler/src/runtime.rs
@@ -1,4 +1,3 @@
-use crate::extension::Plugins;
 use crate::{Compiler, Context, Target};
 
 use std::convert::From;
@@ -6,16 +5,14 @@ use std::convert::From;
 #[derive(Clone)]
 pub struct Runtime {
 	pub context: Context,
-	pub workspace: Target,
-	pub plugins: Plugins,
+	pub target: Target,
 }
 
 impl From<&Compiler> for Runtime {
 	fn from(compiler: &Compiler) -> Self {
 		Runtime {
 			context: compiler.context.clone(),
-			workspace: compiler.workspace.clone(),
-			plugins: compiler.plugins.clone(),
+			target: compiler.target.clone(),
 		}
 	}
 }