+++ crates/kct_package/src/error.rs
@@ -4,13 +4,13 @@ use thiserror::Error;
 pub enum Error {
 	#[error("Package doesn't have a main template")]
 	NoMain,
-	#[error("Package is neither directory nor a .tgz")]
+	#[error("Package is a directory")]
 	InvalidFormat,
 	#[error("Missing package file")]
 	NoSpec,
 	#[error("Invalid package file")]
 	InvalidSpec,
-	#[error("No schema file to validate your input")]
+	#[error("No schema file to validate your example")]
 	NoSchema,
 	#[error("Invalid schema file")]
 	InvalidSchema,
@@ -18,14 +18,6 @@ pub enum Error {
 	NoExample,
 	#[error("Invalid example file")]
 	InvalidExample,
-	#[error("No input was provided")]
-	NoInput,
-	#[error("The input provided doesn't match the schema")]
-	InvalidInput,
-	#[error("An error happened while rendering your templates: {0}")]
-	RenderIssue(String),
-	#[error("Your template couldn't be parsed as JSON")]
-	InvalidOutput,
 }
 
 pub type Result<T> = std::result::Result<T, Error>;